support adding neutron networks to k8s containers
create network attachment definitions with empty ipams on the
default namespace for each baremetal network interface on the
worker node with following name: <worker-name>.<interface-name>.
Then during container creation, you can specify networks to be added
to the container up to the number of baremetal interfaces. By default,
macvlan bridge is assumed as the cni. Multus must be installed on
k8s as well. Add labels in the following format during container
creation to specify which interface to connect to which network
and you can specify the ip as well. If you dont, dhcp is assumed:
e.g.
	networks.1.interface=enp4s0np0,networks.1.ip=192.168.100.200/24
	networks.2.interface=enp5s0np1